+++ lib/src/models/stories/details/story_details.dart
@@ -1,7 +1,0 @@
-import '../../../../insta_extractor.dart';
-
-class StoryDetails {
-  final List<Tray> trays;
-  final Owner owner;
-  StoryDetails(this.owner, this.trays);
-}