+++ lib/src/utils/api_utils.dart
@@ -28,9 +28,9 @@ abstract class ApiUtils {
   static const String imageVersions2 = "image_versions2";
   static const String profilePicUrl = "profile_pic_url";
 
-  static const String USERAGENT =
-      "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4577.82 Safari/537.36";
+  static const String USERAGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4577.82 Safari/537.36";
 
-  static const String STORY_USERAGENT =
-      "Instagram 9.5.2 (iPhone7,2; iPhone OS 9_3_3; en_US; en-US; scale=2.00; 750x1334) AppleWebKit/420+";
+  static const String STORY_USERAGENT = "Instagram 9.5.2 (iPhone7,2; iPhone OS 9_3_3; en_US; en-US; scale=2.00; 750x1334) AppleWebKit/420+";
+
+  static const ENDPOINT = 'https://instaapigram.herokuapp.com/media/';
 }