+++ lib/src/utils/link_utils.dart
@@ -17,13 +17,13 @@ abstract class LinkUtils {
 
 abstract class ParseLink {
   ///Returns the downloadable instagram link
-  ///Format: https://instagram.com/p/abc?__a=1 (sample)
+  ///Format: https://instagram.com/p/abc/ (sample)
   static String instagram(String url) {
     if (url.contains("stories")) {
       var matches = RegExp(LinkUtils.kInstaStoryPattern).firstMatch(url);
       return "${matches!.namedGroup("required")}";
     }
     var matches = RegExp(LinkUtils.kInstagramPattern).firstMatch(url);
-    return "${matches!.namedGroup("required")}?__a=1";
+    return "${matches!.namedGroup("required")}";
   }
 }