+++ CHANGELOG.md
@@ -9,3 +9,8 @@
 ## 0.0.8
 
 * Updated Story response data
+
+## 1.0.0
+* Moved from sharedPreferences to Hive Flutter
+* New api added
+* New Unified Model added **Insta**